+++ src/main/java/com/tarea/samsungshop/domain/repository/DetallePedidoRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.tarea.samsungshop.domain.model.DetallePedido;
 
 @Repository
-public interface DetallePedidoRepository extends JpaRepository<DetallePedido, Long> {
+public interface DetallePedidoRepository extends JpaRepository<DetallePedido, Integer> {
 }
